Rename content to value, implement zero-arg call, move to new files.
